+++ src/physics/operators/nonlinear_solid_operators.cpp
@@ -7,7 +7,7 @@
 #include "physics/operators/nonlinear_solid_operators.hpp"
 
 #include "infrastructure/logger.hpp"
-#include "math/expr_template_ops.hpp"
+#include "numerics/expr_template_ops.hpp"
 
 namespace serac {
