+++ tests/serac_component_bc.cpp
@@ -9,8 +9,8 @@
 #include <fstream>
 
 #include "coefficients/stdfunction_coefficient.hpp"
-#include "math/mesh_utils.hpp"
 #include "mfem.hpp"
+#include "numerics/mesh_utils.hpp"
 #include "physics/nonlinear_solid_solver.hpp"
 #include "serac_config.hpp"
