+++ tests/serac_dtor.cpp
@@ -10,8 +10,8 @@
 #include <fstream>
 #include <memory>
 
-#include "math/mesh_utils.hpp"
 #include "mfem.hpp"
+#include "numerics/mesh_utils.hpp"
 #include "physics/thermal_solver.hpp"
 #include "serac_config.hpp"
