+++ tests/serac_mesh.cpp
@@ -8,8 +8,8 @@
 
 #include <fstream>
 
-#include "math/mesh_utils.hpp"
 #include "mfem.hpp"
+#include "numerics/mesh_utils.hpp"
 #include "physics/elasticity_solver.hpp"
 #include "serac_config.hpp"
