+++ tests/serac_quasistatic_solver.cpp
@@ -9,8 +9,8 @@
 #include <fstream>
 
 #include "coefficients/loading_functions.hpp"
-#include "math/mesh_utils.hpp"
 #include "mfem.hpp"
+#include "numerics/mesh_utils.hpp"
 #include "physics/nonlinear_solid_solver.hpp"
 #include "serac_config.hpp"
